update UI to support gridded reports
