+++ apps/fishing-map/public/locales/source/translations.json
@@ -209,6 +209,7 @@
     "activityAction": "Download visible activity layers for this area",
     "activityLogin": "Register and login to download activity (free, 2 minutes)",
     "area": "Area",
+    "byVessel": "By vessel",
     "daily": "Daily",
     "datasetsNotAllowed": "You don't have permissions to download the following datasets:",
     "format": "Format",
@@ -216,6 +217,8 @@
     "highResNotAvailable": "Your area is too big",
     "monthly": "Monthly",
     "monthlyNotAvailable": "Your time range is shorter than 1 month",
+    "fullTimeRange": "Full time range",
+    "gridded": "Gridded",
     "spatialResolution": "Spatial resolution",
     "temporalResolution": "Temporal resolution",
     "timeRange": "Time Range",
